Rename libceed folder to ceed.
