+++ fem/ceed/operator.cpp
